big changes in RAGE::SDL design... this is likely not to work, but still it builds.
The Engine is not related to the VideoSurface anymore. Instead Window takes care of calling it.
This simplify the design a lot, and allow us to get rid of GLSurface, which is useless from now on.

The Window Class has also been a bit simplified :)

Lost of testing to do, and cleaning in progress...