Project 0 2D v0.02:
Externalization of characters movement check according to ground type in order to allow a better instantiation later when we will have more than one character.

Create two monsters (skeleton & worm) derivate from monster_base and set the movement rules according to the ground type.


Lots of work in trying to transform the monster factory to a template class... still not working for the moment (linking problem)