Project 0 2D v0.02:
Playing with cmake: 
-> good datas copied in order to have debug working
-> now include all sdlut dependencies AND opengl32.libpng12.dll
->project is now created in the binary dir and not in the binary/src dir
-> other tweaks

Anyway Finally cmake file works!!

Only 2 things to tweaks : find a way to ask the user for SDLut's CMAKE_MODULE_PATH and for the file that define SDLut's dependences.

NB: there is also a bug in copying the Arrow.bmp file with the cmake CONFIGURE_FILE() cmd: it destroy the final file 8k instead of 13k. its a cmake bug!